Added authentication to cart and shows cart according to logged user

+++ src/components/Cart/RemoveCart.js
@@ -1,14 +1,19 @@
 import {db} from '../Firebase/firebase'
 import { deleteDoc, doc } from 'firebase/firestore'
+import { getAuth } from 'firebase/auth';
 
 export const removeCartItem = async(id, setItems) => {
-    try{
-        const docItem = doc(db, 'Cart', id)
-        await deleteDoc(docItem)
+    const auth = getAuth();
+    const user = auth.currentUser;
+    if(user){
+        try{
+            const docItem = doc(db, 'Cart', user.uid,'CartItems',id)
+            await deleteDoc(docItem)
 
-        setItems((prev) => prev.filter((item) => item.id !==id))
-        console.log("Item removed successfully")
-    }catch(error){
-        console.error("Item not removed",error)
+            setItems((prev) => prev.filter((item) => item.id !==id))
+            console.log("Item removed successfully")
+        }catch(error){
+            console.error("Item not removed",error)
+        }
     }
 }
